Change 1-3 = Item class Change 4-6 = Room class Change 7-9 = Player class Change 10-11 = Adventure class Change 12 = Map class Change 13 = Extra fix Player class
1. Item class:
Added 'Item' class with attributes 'name' and 'description'.,

2. Item Class:
Added a constructor for the class and getter methods.

3. Item class:
Added toString method to the Item class to use the 'name' of the item as a simple representation

4. Room class:
removed the outcommented part of the attribute 'private List<Item> items;' Initialised items as a list by removing outcommented part of contructor from previous push.

5. Room class:
Added method to add and remove items from the room.

6. Room class:
Added list for items and getItemString. Also added the getItemString to the room description

7. Player class:
Created 'player' class

8. Player class:
Implemented 'addItem' and 'removeItem' methods to the class

9. Player class:
Added method List<> getInventory and  getInventoryString as a string to return using stringbuilder method.

10. Adventure class:
Created new player class and initialised it

11. Adventure class:
Implemented ui.printMessage(player.getInventoryString() to show the players items

12. Map class:
Added weapons to room 3, 6 and 8.

13. Player class:
Changed the String at getInventoryString to insert ',' and '&' depending on the index between the inventory objects